Add a new empty class, Charge
This new class will be the delegate class for a controller class,
OmisePaymentModuleFrontController. It will be reduced the dependencies
of the controller class and it will be used to interface with Omise-PHP

The directory that used to store the new class has followed the style of
PrestaShop. The new class has been stored in the new directory, classes.