Change visibility of Charge functions to protected
Currently, it has no any class or function that will be use the
functions within class, Charge.

For this situation, the appropriate visibility for the functions is
private but due to this project is an open source project, it may be
possible that other developers want to extend this class.

So, the more appropriate visibility is protected.