Add a condition to check the defined PS module dir
Add a condition to check the defined PrestaShop module directory for
class, Charge.

The main intention for this condition is to cut the dependencies which
it allows to create mock object for the dependencies easier in part of
unit testing.

If this condition has not been added, in unit testing, it needs to define a
constant that has a valid value for the dependencies directory path
otherwise when run the unit testing, it will be error because the system
can not found the valid path for the dependencies.

In the situation that a valid constant has been defined in unit testing,
it cause to the dependencies have been executed or defined and create
mock will be harder because the actual dependencies have been executed
or defined.