Fix an error when run all unit tests
When run all unit tests with the command,

$ phpunit

the error is

Mockery\Exception\RuntimeException: Could not load mock \Setting, class
already exists

The error occurred because it has been defined mocked class, Setting, 2
places.

1. OmiseTest
2. ChargeTest

To solve this error, it can be solved by add 2 annotations of PHPUnit to
separate the process and memory to prevent the duplcated class defining.

These annotations have a side effect making the test run slower.

So, for the next step, it need to refactor the unit tests to make the
unit test run faster.

Note:
This error is not occurred when run unit test with one test class.

For examples:

$ phpunit --filter OmiseTest

Or

$ phpunit --filter ChargeTest