Add a new empty class, PaymentOrder
This new class will be the delegate class for a controller class,
OmisePaymentModuleFrontController. The delegate class will interfaces
with PrestaShop core objects and functions in part of order.

This new class has been stored in the directory, classes, which the
style of directory structure is similart to the PrestaShop's style.

Note:
This new delegate class can not be named `Order` because PrestaShop
already has a class Order.