Add a unit test for a class, PaymentOrder
The base class, PaymentOrder, has a hard dependency with the class,
Customer. In the class, PaymentOrder, it has a command to instantiate a
class Customer.

$customer = new Customer($cart->id_customer);

And the base class, PaymentOrder, has the access to the public property
of the class, Customer.

return $customer->secure_key;

This situation, access the public property of the hard dependency class,
can not be mocked with PHPUnit and Mockery because:

- PHPUnit can not mock the hard dependency (create new instance)
- Although, Mockery can mock the hard dependency but access to the
  public property of the hard dependency class, when run unit test with
  PHPUnit the error is the public propery has not been defined.

To cope with this situation, it can be:

1. Modify the base class, PaymentOrder, by implement the dependency
  injection for a class, Customer

or

2. Create stub class for the class, Customer, in part of unit testing

Although, the implement of dependency injection is preferred but a
disadvantage of dependency injection is the complexity. The base class
will be modified by add a class property and a function to inject the
dependency. This modification has the main intention for unit testing
only is quite not worth.

So, this situation the unit testing using the create stub class
strategy.