Implement dependency injection for a property
Modify a payment controller class, OmisePaymentModuleFrontController.
Implement the dependency injection that depend to the class, PaymentOrder.

This modification has main intention for the easier to create mock
object in part of unit testing and also it is better to make the function
has source code for business logic only.

Although, it has a remaining hard dependency with a class, OmiseCharge.
It can be refactored for the next time.