Change return type of Charge.create() to itself
Modify a function, Charge.create(). Change the return type from
OmiseCharge to itself.

This change will allow the mechanism to wrap the knowledge or
information about the Omise charge within the class, Charge and this
change is the first step to provide the information such as error
message or failure to the caller class.