Implement a hard dependency to class, PaymentOrder
Modify a class, OmisePaymentModuleFrontController. Remove the dependency
injection of a field, payment_order.

Before change (this commit), the dependency, PaymentOrder, has been
injected via the setter function at the constructor of a class
OmisePaymentModuleFrontController and in the constructor, it need to
execute the parent constructor.

This execute has the impact to the front-end display which make the
display is incorrect.

This problem will occur if the display using the function setTemplate().