Separate the process to run ChargeTest
When run all unit test with a command, phpunit, a runtime exception will
be occurred. The runtime exception is Mockery\Exception\RuntimeException: Could
not load mock \Setting, class already exists.

To solve this problem, it can be solved by separate the process to
run each unit test by adding 2 PHPUnit annotations.

1. @runTestsInSeparateProcesses
2. @preserveGlobalState disabled

But those annotations impact to the time when run unit test.

Another solution is refactor the unit test but this refactor may be
affect to the base source code need to be modified.