Add a parameter for Charge.create()
Modify a function, Charge.create(). Add a parameter, card token, that
will be passed from the controller class.

This change has followed the style of PrestaShop source code that the
controller will receive the parameter from client side and pass it through
process.

Although, some class of PrestaShop such as objec model class has
dierctly receive the parameter from client side. The parameter does not
passed from the controller but it was only a few class when compare with receive the
parameter at controller class.