Remove a function that used to handle capture
Remove a function, Charge.getCapture(), that used to define the value
for a parameter, capture, when create Omise charge.

Currently, it has only create Omise charge by auto capture.

To create Omise charge by manual capture, it has not been developed.

So, a function that used to handle the capture parameter when create
Omise charge will be developed along with a manual capture feature.