Add prefix to the keys that used to store setting
Modify a class, Setting. Add the prefix, omise_, to all key that used to
store the setting configuration.

This change is for prevent the duplicate key problem with core
PrestaShop or others third-party module.

The prefix is added to the keys that used to store the setting only.
The keys that used to get value from view (client side) are remain the same.