Add a function to delete all setting from database
Modify a class, Setting. Add a function that used to delete all module
setting from database.

This function is used when uninstall the module.

Before change, when uninstall the module, the setting values are not
deleted. It may be the problem that the sensitive data, Omise public key
and Omise secret key can be access ithout permission or by mistake in
spite of the module has been uninstalled.