Remove a link, Back to payment, when payment error
Modify a view, payment-error.tpl. Remove a link, Back to payment.

This view has been shared with normal payment and 3-D Secure payment.

For the normal payment, display the link, Back to payment, is no
problem. The system can go back to the payment step correctly.

But for the 3-D Secure payment, it has a step to save an order with
in progress order state. This order save step is for generate a
PrestaShop order reference. The reference used to send to Omise API
for create 3-D Secure charge. But when create 3-D Secure charge is fail
or error, the system will display the view, payment-error.tpl,
as well but the link, Back to payment, will work not properly because
when the order has been save the cart information has been cleared.
So, when go back to the payment step, the system will display the error
message.

Although, this change has the impact on the normal payment but the order
steps panel can be an alternative for the customer to go back to the
previous step.