Redirect to first checkout step, if cart invalid
Modify functions, OmisePaymentModuleFrontController.postProcess() and
OmiseThreeDomainSecurePaymentModuleFrontController.postProcess().

Add a function to validate the cart information. If the cart information
is not valid, the system will redirect the page to the first checkout step.

This change prevents the problem in situation that the process of create
Omise charge (e.g. the URL, /module/omise/payment) has been accessed
directly instead of submit from the checkout form at the client side or
accessed without any cart information.

The problem is the unnecessary request has been sent through the Omise
API instead of this condition can be checked and prevented at the server
side.