Remove 2 assigned variables that displayed form
Remove 2 assigned variables that used to display at card payment form.
The removed variables are `action` and `title`.

PrestaShop 1.7 has changed the mechanism to submit payment form.
It will generated a submit button for all payment options.

So, each payment option will be changed to override the action of the
PrestaShop 1.7. instead of has it own action.

The `title` at the top of the payment form has changed to be displayed
at the payment option instead. So, display the title at the top of
payment form is unnecessary and duplicated.