Re-add missing condition, Is Omise.js accessible?
Re-add a missing JavaScript condition. This condition is lost during
the modification of previous commit.

This condition is check whether the external library, Omise.js,
is accessible. If Omise.js is not accessible, the error message will be
displayed. The reason that Omise.js is not accessible, may be from the
slow connection or timeout.

The Omise.js is an important library that used to create Omise card
token at the client side.

Without this condition, the customer (payer) can not clearly know the
problem when submit payment with Omise card payment. The error will be
not displayed visually. It will be displayed at the console of web browser.
That is not easy for customer to know the problem.