Prevent error notice at setting page on debug mode
In the setting page, it has a variable, `confirmation`. This variable is
used for display the confirmation message when the merchant saved the
setting.

If merchant enabled the debug mode and open the setting page, the system
displays the error message popup, Undefined index: confirmation.

This problem is occurred because at the time that merchant open the setting
page and does not save the setting. The variable, `confirmation`,
has not been assigned any value from server side to display any message.

This problem is not occurred in normal mode.