Specify Omise API version to be 2015-11-17
The Omise API version, 2015-11-17, is compatible with the current
implementation of create internet banking charge.

If the Omise API version has not been specified, when create internet
banking charge with the Omise API version that incompatible, the error
is occurred.

Currently, the new version of Omise API is 2017-11-02. This new version
is not compatible in part of create internet banking charge.

So, the new merchant that registered the Omise account since 02 Nov
2017 or the old merchant that update Omise API version of their account,
will have the problem as mentioned above.