Partially add unit test for a controller class
Partially add unit test cases for a controller class,
OmiseReturnModuleFrontController.

The unit test for this class can not be additionally added because this
class has a hard dependency (new keyword) with a class, Order, and access
to many public properties of Order class (e.g. $this->order->reference).

This condition makes the mocking public property for hard dependency
difficult.

Another solution is adding a setter for the $this->order but this
solution needs to modify the actual code that it is not requires in this
time.