Refactor, change a param name of PaymentOrder.save
Modify a name of parameter of a function, PaymentOrder.save().
Change a name of parameter from $order_state to $id_order_state.

Make the name consistent with PrestaShop. An example from PrestaShop
function is PaymentModule.validateOrder().

A parameter that PaymentOrder.save() required is ID of order status.
Using $order_state (before change) can be confused that it is the
instance of OrderState class.