Check module of saved order when return from Omise
Check the module name of saved order when return from Omise API.

According to the payment method of saved order has been dynamically
saved follows the payment method that customer selected. Such as card
payment or internet banking payment.

Therefor, payment method of saved order (order->payment) is variable.
It can not be used to check whether the order is the order that made
payment with Omise.

So, using module name of saved order (order->module) to check with
Omise.MODULE_NAME is properly.