Change flow when creating payment by card
This commit affects normal card payment only. 3DS payment is not
affected.

The objective of this commit is to append PrestaShop order ID to the
metadata of Omise charge.

The flow has been changed to save an order as processing at the first
step and uses that generated order ID to append to the metadata of Omise charge.