Develop a new class, OmiseLogger, for write log
This class wrapped the PrestaShop logger class, Logger. The Logger class
requires some parameters which they are not required or compatibled with
module such as $object_id or $object_type. The data type of $object_id is
int but the data type of Omise object id is string or $object_type is
the internal PrestaShop object, etc.

So, the wrapper class, OmiseLogger, has reduced the parameters that it is
not used by provides the necessary parameters, $message and $severity.
And $severity is optional the default value of $severify is info log
level.

The same log message can be added to database because OmiseLogger has
set a parameter, $allow_duplicate, to be true. It has the intent that
the same log message but differnt time can be added to database.