Add a unit test for OmiseEventHandler
The unit test in this commit has only one test case because in actual
class, OmiseEventHandler, in has a static class attribute dependency,
OmiseEventChargeComplete::KEY, and a hard dependency (new Keyword),
new OmiseEventChargeComplete(), with the same class. This situation is
hard to create mock for unit testing.

So, a unit test case in this commit that test class, OmiseEventHandler,
is the test case that across the dependency that hard to mock.