Remain status, processing, if charge is not paid
Remains the PrestaShop order status to be Processing in progress, if
Omise charge is not paid.

This commit has an objective to solve the possible problem of offsite
payment such as internet banking payment that the notification from bank
server to Omise server is delayed and Omise can not knows or updates the
charge status.

For the internet banking payment, when the payer completed their payment
on bank website and bank redirects payer back to PrestaShop site.

In this step, it is possible that the notification from bank server to
Omise server can be deplayed or slower than the redirection of payer.
This situation causes the charge status is pending and charge is not
paid because Omise does not know the status of payment.

So, it is safer that if the charge is not paid the PrestaShop order
status still Processing in progress.

Note:
The implementation of webhooks to prevent this problem by correctly
synchronize the charge information between PrestaShop order and Omise
charge has been completed in the previous commits.