Correctly handle the Omise charge status of paid
When the payer has been redirected back to PrestaShop, the module, Omise
PrestaShop, will checks the information of Omise charge by request the
information via Omise API.

Before change:
If Omise charge is NOT paid, the empty error message on the order
confirmation page will be displayed. This result is not correct and properly
because in this step, the PrestaShop order is valid and Omise charge is
not failed. The status of Omise charge is not paid. It does not mean
that the payment is error.

After change:
The success order confirmation page is display normally. Regardless of
Omise charge status is paid or unpaid.

The main objective of before change and after change still the same. If the
Omise charge is paid, the PrestaShop order status will be updated
to be success (Payment accepted).