Remove a condition checking PS order status
Remove a condition that checking current PrestaShop order status,
when redirect payer back to PrestaShop site.

This commit has an objective to solve a problem of internet banking
payment. The problem occurs in the situation that;

the redirection of payer back to PrestaShop site is slower than the
payment result notification from bank to Omise and Omise charge of this
payment is failed.

This situation causes the invalid result message on PrestaShop order
confirmation page.

Note:
Due to the offsite payment that has the redirection flow such as
internet banking payment, including the implementation of Omise
webhooks, it is possible that;

- Omise charge is failed.
- The notification from bank to Omise was passed or forwarded to the
PrestaShop site via Omise webhooks FASTER THAN the redirection of payer
to PrestaShop.
- Omise PrestaShop handles this webhooks event by change the PrestaShop
order status to be Canceled.

So, this condition or situation causes the invalid result message on
PrestaShop order confirmation page.