Use CSS to style the text label of checkbox
Add the HTML tag, <span>, to the text label of checkbox, Remember this
card.

Before change, to add space between checkbox and its text label, it uses
a whitespace. This whitespace does not make the proper style when
compare with using the CSS.

After added the HTML tag, <span>, the CSS rule has been applied and it
made the display of label and checkbox properly. The label and checkbox
are aligned horizontally and space between them has been specified
by using CSS. Including for the customization, for an example, the
third-party theme, it is more easier customize by using the CSS rule.

Note:
This change is using the default theme of PrestaShop 1.7.