Add a upgrade function create table omise_customer
Add a upgrade function to create a database table, omise_customer.

Without this upgrade function, the installation of Omise PrestaShop new
version, 1.5, on the installed version such as 1.4, the database table,
omise_customer, will be NOT created.

But if install the Omise PrestaShop new version, 1.5, on the PrestaShop
that has not installed Omise PrestaShop, this problem is not happened.
The database table, omise_customer, will be installed successfully.

Reference:
http://doc.prestashop.com/display/PS17/Enabling+the+Auto-Update